+++ src/hardware_language_abstraction/hardware_language_abstraction_teensy35.h
@@ -158,7 +158,7 @@ void set_stepper_int_function(void (*f)());
 
 //Complete setup (long)
 
-void setup_stepper_interrupt(void (*function)(void), uint16_t period_us);
+void setup_stepper_interrupt(void (*function)(void), uint32_t period_us);
 
 
 //Complete clean
